LPS-50046 Move to same package as ServiceContext
